+++ test/fixtures/classes.expected.js
@@ -84,3 +84,20 @@ class Josie extends Error {
 }
 
 class Gizmo {}
+
+class Hannah extends Error {
+  constructor(Tucker = 'Tucker') {
+    super(Charlie);
+    this.name = this.constructor.name;
+    this.message = Charlie;
+
+    if (typeof Error.captureStackTrace === 'function') {
+      Error.captureStackTrace(this, this.constructor);
+    } else {
+      this.stack = new Error(Charlie).stack;
+    }
+
+    console.log(Tucker);
+  }
+
+}